+++ src/components/Tiptap/index.tsx
@@ -4,11 +4,9 @@ import Underline from "@tiptap/extension-underline";
 import Image from "@tiptap/extension-image";
 import Link from "@tiptap/extension-link";
 import { ValidationError } from "yup";
-import AddPhotoAlternateIcon from "@mui/icons-material/AddPhotoAlternate";
 
-// models
-import { imageFormSchema } from "src/models/forms/imageForm";
-import { useEffect } from "react";
+// components
+import MenuBar from "./MenuBar";
 
 type Props = {
   value: string;
@@ -44,44 +42,13 @@ const Tiptap = (props: Props) => {
     [!!value]
   );
 
-  const TiptapMenuBar = (props: {}) => {
-    const handleClickImageButton = () => {
-      if (typeof document === undefined) return;
-
-      const input = document.createElement("input");
-      input.setAttribute("type", "file");
-      input.setAttribute("accept", "image/*");
-      input.click();
-      input.onchange = async () => {
-        const file = input?.files?.[0];
-
-        if (!file) return;
-
-        await imageFormSchema
-          .validate({ image: file })
-          .then(({ image }) => {
-            onSubmitImage(image, editor);
-          })
-          .catch((error) => {
-            onErrorSubmitImage(error);
-          });
-      };
-    };
-
-    return (
-      <>
-        <div className="menu-bar">
-          <button onClick={handleClickImageButton} type="button">
-            <AddPhotoAlternateIcon fontSize="large" />
-          </button>
-        </div>
-      </>
-    );
-  };
-
   return (
     <>
-      <TiptapMenuBar />
+      <MenuBar
+        editor={editor}
+        onSubmitImage={onSubmitImage}
+        onErrorSubmitImage={onErrorSubmitImage}
+      />
       <EditorContent editor={editor} className="editor" />
     </>
   );
